improve headless comp names
