+++ dev/example_project/example_project/templates/simple_cotton.html
@@ -1,3 +1,1 @@
-{#{% for d in data %}<c-benchmarks.cotton-include />{% endfor %}#}
-{% load cache %}
-{% for d in data %}{% comp benchmarks.cotton-include %}{% endcomp %}{% endfor %}
+{% for d in data %}{% c benchmarks.cotton-include %}{% endc %}{% endfor %}
