--- dev/example_project/v2/apps.py
@@ -1,6 +1,0 @@
-from django.apps import AppConfig
-
-
-class V2Config(AppConfig):
-    default_auto_field = 'django.db.models.BigAutoField'
-    name = 'v2'
