--- dev/example_project/v2/templates/cotton/button.html
@@ -1,3 +1,0 @@
-<button class="{{ attrs.class|default:'btn' }}">
-    {{ slot }}
-</button>
