--- dev/example_project/v2/templates/cotton/card.html
@@ -1,21 +1,0 @@
-{% cvars title var_override="default!" %}
-
-<div style="margin: 5px 20px; border: 1px solid #ccc; border-radius: 5px">
-{#    {% querystring %}#}
-    <h2>{{ title }}</h2>
-    <div class="card-content">
-        {{ slot }}
-
-        <p>attrs: {{ attrs }}</p>
-        <p>attrs_dict: {{ attrs.dict }}</p>
-        <p>var_override?: {{ var_override }}</p>
-    </div>
-    {% if footer %}
-        <div style="font-weight: bold">
-            {{ footer }}
-        </div>
-    {% endif %}
-
-</div>
-
-{%  endcvars %}
