+++ docs/docs_project/docs_project/settings.py
@@ -77,7 +77,6 @@ TEMPLATES = [
                 "docs_project.templatetags.force_escape",
                 "docs_project.templatetags.custom_tags",
                 "heroicons.templatetags.heroicons",
-                "django_cotton.templatetags.tags",
                 "django_cotton.templatetags.cotton",
             ],
         },
